chore: reorganize window and item code
