+++ static/app/js/items/llm_service_model_item.js
