+++ static/app/js/items/prompt_template_item.js
