+++ static/app/js/items/search_result_item.js
