+++ static/app/js/items/segment_item.js
